locale: Update locale configs
